refactoring peak detector.
