+++ tridesclous/peeler_engine_geometry.py
@@ -11,7 +11,7 @@ from concurrent.futures import ThreadPoolExecutor
 
 
 
-from .peeler_engine_classic import PeelerEngineClassic
+from .peeler_engine_base import PeelerEngineGeneric
 
 from .peeler_tools import *
 from .peeler_tools import _dtype_spike
@@ -38,19 +38,16 @@ except ImportError:
 import matplotlib.pyplot as plt
 
 
-class PeelerEngineGeometry(PeelerEngineClassic):
-    pass
+class PeelerEngineGeometry(PeelerEngineGeneric):
     def change_params(self, adjacency_radius_um=200, **kargs):
-        PeelerEngineClassic.change_params(self, **kargs)
+        PeelerEngineGeneric.change_params(self, **kargs)
         
         self.adjacency_radius_um = adjacency_radius_um
         
         assert self.use_sparse_template
-        
-        
 
     def initialize_before_each_segment(self, **kargs):
-        PeelerEngineClassic.initialize_before_each_segment(self, **kargs)
+        PeelerEngineGeneric.initialize_before_each_segment(self, **kargs)
 
         p = dict(self.catalogue['peak_detector_params'])
         _ = p.pop('peakdetector_engine', 'numpy')
@@ -71,131 +68,6 @@ class PeelerEngineGeometry(PeelerEngineClassic):
             nearest, = np.nonzero(d[c, :]<self.adjacency_radius_um)
             self.channels_adjacency[c] = nearest
             #~ print(c, nearest)
-        
-    
-    
-#~ self.sparse_mask
-
-    def process_one_chunk(self,  pos, sigs_chunk):
-        print('process_one_chunk', pos)
-        t1 = time.perf_counter()
-        abs_head_index, preprocessed_chunk = self.signalpreprocessor.process_data(pos, sigs_chunk)
-        
-        #shift rsiruals buffer and put the new one on right side
-        t1 = time.perf_counter()
-        fifo_roll_size = self.fifo_residuals.shape[0]-preprocessed_chunk.shape[0]
-        if fifo_roll_size>0 and fifo_roll_size!=self.fifo_residuals.shape[0]:
-            self.fifo_residuals[:fifo_roll_size,:] = self.fifo_residuals[-fifo_roll_size:,:]
-            self.fifo_residuals[fifo_roll_size:,:] = preprocessed_chunk
-
-        # relation between inside chunk index and abs index
-        shift = abs_head_index - self.fifo_residuals.shape[0]
-        
-        # negative mask 1: not tested 0: already tested
-        self.mask_not_already_tested[:] = True
-        
-        #~ self.local_peaks_mask = self.peakdetector.get_mask_peaks_in_chunk(self.fifo_residuals)
-        self.local_peaks_mask = self.peakdetector.get_mask_peaks_in_chunk(self.fifo_residuals)
-        
-        
-        good_spikes = []
-        
-        n_loop = 0
-        t3 = time.perf_counter()
-        while True:
-            #~ print('peeler level +1')
-            nb_good_spike = 0
-            peak_ind, peak_chan = self.select_next_peak()
-            
-            
-            #~ print('start inner loop')
-            while peak_ind != LABEL_NO_MORE_PEAK:
-            
-                t1 = time.perf_counter()
-                spike = self.classify_and_align_next_spike(peak_ind, peak_chan)
-                #~ print(spike)
-                
-                if spike.cluster_label == LABEL_NO_MORE_PEAK:
-                    break
-                
-                if (spike.cluster_label >=0):
-                    #~ good_spikes.append(np.array([spike], dtype=_dtype_spike))
-                    good_spikes.append(spike)
-                    nb_good_spike+=1
-                    
-                    # remove from residulals
-                    self.on_accepted_spike(spike)
-                    
-                    self.mask_not_already_tested[peak_ind - self.n_span, peak_chan] = False # this save lot of time
-                else:
-                    # set this peak_ind as already tested
-                    self.mask_not_already_tested[peak_ind - self.n_span, peak_chan] = False
-
-                peak_ind, peak_chan = self.select_next_peak()
-                #~ print(peak_ind, peak_chan)
-                
-                
-                #~ # debug
-                n_loop +=1 
-                
-            if nb_good_spike == 0:
-                #~ print('break main loop')
-                break
-            else:
-                
-                t1 = time.perf_counter()
-                for spike in good_spikes[-nb_good_spike:]:
-                    peak_ind = spike.index
-                    # TODO here make enlarge a bit with maximum_jitter_shift
-                    #~ sl1 = slice(peak_ind + self.n_left - 1 - self.n_span, peak_ind + self.n_right + 1 + self.n_span)
-                    #~ sl2 = slice(peak_ind + self.n_left - 1 - self.n_span, peak_ind + self.n_right + 1- self.n_span)
-                    #~ self.local_peaks_mask[sl2] = self.peakdetector.get_mask_peaks_in_chunk(self.fifo_residuals[sl1, :])
-                    #TODO
-                    self.local_peaks_mask = self.peakdetector.get_mask_peaks_in_chunk(self.fifo_residuals)
-                    
-                    # set neighboor untested
-                    # TODO more efficient !!!!!!!!!
-                    cluster_idx = self.catalogue['label_to_index'][spike.cluster_label]
-                    mask = self.catalogue['sparse_mask'][cluster_idx, :]
-                    for c in np.nonzero(mask)[0]:
-                        #~ print(c)
-                        self.mask_not_already_tested[peak_ind - self.peak_width - self.n_span:peak_ind + self.peak_width - self.n_span, c] = True
-                #~ t2 = time.perf_counter()
-                #~ print('  update mask', (t2-t1)*1000)
-
-
-        #~ t4 = time.perf_counter()
-        #~ print('LOOP classify_and_align_one_spike', (t4-t3)*1000)
-        #~ print('nb_good_spike', len(good_spikes), 'n_loop', n_loop)
-        
-        nolabel_indexes, chan_indexes = np.nonzero(~self.mask_not_already_tested)
-        nolabel_indexes += self.n_span
-        nolabel_indexes = nolabel_indexes[nolabel_indexes<(self.chunksize+self.n_span)]
-        bad_spikes = np.zeros(nolabel_indexes.shape[0], dtype=_dtype_spike)
-        bad_spikes['index'] = nolabel_indexes + shift
-        bad_spikes['cluster_label'] = LABEL_UNCLASSIFIED
-        
-        if len(good_spikes)>0:
-            # TODO remove from peak the very begining of the signal because of border filtering effects
-            
-            good_spikes = np.array(good_spikes, dtype=_dtype_spike)
-            good_spikes['index'] += shift
-            near_border = (good_spikes['index'] - shift)>=(self.chunksize+self.n_span)
-            near_border_good_spikes = good_spikes[near_border].copy()
-            good_spikes = good_spikes[~near_border]
-
-            all_spikes = np.concatenate([good_spikes] + [bad_spikes] + self.near_border_good_spikes)
-            self.near_border_good_spikes = [near_border_good_spikes] # for next chunk
-        else:
-            all_spikes = np.concatenate([bad_spikes] + self.near_border_good_spikes)
-            self.near_border_good_spikes = []
-        
-        # all_spikes = all_spikes[np.argsort(all_spikes['index'])]
-        all_spikes = all_spikes.take(np.argsort(all_spikes['index']))
-        self.total_spike += all_spikes.size
-        
-        return abs_head_index, preprocessed_chunk, self.total_spike, all_spikes
-
 
     def select_next_peak(self):
         #~ print('select_next_peak')
@@ -210,153 +82,48 @@ class PeelerEngineGeometry(PeelerEngineClassic):
             peak_ind = local_peaks_indexes[ind]
             chan_ind = chan_indexes[ind]
             return peak_ind, chan_ind
-            #~ return local_peaks_indexes[0]
         else:
             return LABEL_NO_MORE_PEAK, None
 
-    def classify_and_align_next_spike(self, proposed_peak_ind, chan_ind):
-        # left_ind is the waveform left border
-        left_ind = proposed_peak_ind + self.n_left
+    def on_accepted_spike(self, spike):
+        # remove spike prediction from fifo residuals
+        left_ind = spike.index + self.n_left
+        cluster_idx = self.catalogue['label_to_index'][spike.cluster_label]
+        pos, pred = make_prediction_one_spike(spike.index, cluster_idx, spike.jitter, self.fifo_residuals.dtype, self.catalogue)
+        self.fifo_residuals[pos:pos+self.peak_width, :] -= pred
         
-        #~ print('classify_and_align_next_spike')
-        #~ print('chan_ind', chan_ind)
-        #~ print(self.catalogue['sparse_mask'].shape)
-        
-        #~ fig, ax = plt.subplots()
-        #~ ax.plot(self.fifo_residuals[:, chan_ind])
-        #~ ax.scatter([proposed_peak_ind], [self.fifo_residuals[proposed_peak_ind, chan_ind]], color='r')
-        #~ plt.show()
-        
-        
-        
-
-        if left_ind+self.peak_width+self.maximum_jitter_shift+1>=self.fifo_residuals.shape[0]:
-            # TODO : remove this because maybe unecessry
-            # too near right limits no label
-            label = LABEL_RIGHT_LIMIT
-            jitter = 0
-        elif left_ind<=self.maximum_jitter_shift:
-            # TODO : remove this because maybe unecessry
-            # too near left limits no label
-            #~ print('     LABEL_LEFT_LIMIT', left_ind)
-            label = LABEL_LEFT_LIMIT
-            jitter = 0
-        elif self.catalogue['centers0'].shape[0]==0:
-            # empty catalogue
-            label  = LABEL_UNCLASSIFIED
-            jitter = 0
-        else:
-            waveform = self.fifo_residuals[left_ind:left_ind+self.peak_width,:]
-            
-            if self.alien_value_threshold is not None and \
-                    np.any((waveform>self.alien_value_threshold) | (waveform<-self.alien_value_threshold)) :
-                label  = LABEL_ALIEN
-                jitter = 0
-            else:
-                
-                #~ t1 = time.perf_counter()
-                #TODO try usewaveform to avoid new buffer ????
-                
-                cluster_idx = self.get_best_template(left_ind, chan_ind)
-                #~ t2 = time.perf_counter()
-                #~ print('    get_best_template', (t2-t1)*1000)
-                
+        # this prevent search peaks in the zone until next "reset_to_not_tested"
+        self.set_already_tested_spike_zone(spike.index, cluster_idx)
 
 
-                
-                
-                #~ t1 = time.perf_counter()
-                jitter = self.estimate_jitter(left_ind, cluster_idx)
-                #~ t2 = time.perf_counter()
-                #~ print('    estimate_jitter', (t2-t1)*1000)
-                
-                #~ t1 = time.perf_counter()
-                ok = self.accept_tempate(left_ind, cluster_idx, jitter)
-                #~ t2 = time.perf_counter()
-                #~ print('    accept_tempate', (t2-t1)*1000)
+    def set_already_tested_spike_zone(self, peak_ind, cluster_idx):
+        mask = self.sparse_mask[cluster_idx, :]
+        for c in np.nonzero(mask)[0]:
+            self.mask_not_already_tested[peak_ind + self.n_left - self.n_span:peak_ind + self.n_right- self.n_span, c] = False
+    
+    def set_already_tested(self, peak_ind, peak_chan):
+        self.mask_not_already_tested[peak_ind - self.n_span, peak_chan] = False
 
-                # DEBUG
-                #~ label = self.catalogue['cluster_labels'][cluster_idx]
-                #~ if label in (5, 8):
-                #~ if label in (10, ):
-                    #~ print('label', label, 'ok', ok, 'jitter', jitter)
-                # END DEBUG                
 
-                if  not ok:
-                    label  = LABEL_UNCLASSIFIED
-                    jitter = 0
-                else:
-                    #~ print('cluster_idx', cluster_idx, 'jitter', jitter)
-                    shift = -int(np.round(jitter))
-                    if (np.abs(jitter) > 0.5) and \
-                            (left_ind+shift+self.peak_width<self.fifo_residuals.shape[0]) and\
-                            ((left_ind + shift) >= 0):
-                        #~ shift = -int(np.round(jitter))
-                        
-                        # debug
-                        #~ new_cluster_idx = self.get_best_template(left_ind+shift)
-                        #~ new_jitter = self.estimate_jitter(left_ind + shift, new_cluster_idx)
-                        #~ ok = self.accept_tempate(left_ind+shift, new_cluster_idx, new_jitter)
-                        # end debug
-                        new_jitter = self.estimate_jitter(left_ind + shift, cluster_idx)
-                        ok = self.accept_tempate(left_ind+shift, cluster_idx, new_jitter)
-                        if ok and np.abs(new_jitter)<np.abs(jitter):
-                            jitter = new_jitter
-                            left_ind += shift
-                            shift = -int(np.round(jitter))
-                            
-                            # debug
-                            #~ if cluster_idx != new_cluster_idx:
-                                #~ print('cluster_idx != new_cluster_idx')
-                            #~ cluster_idx = new_cluster_idx
-                            
-                            
-                    
-                    # ensure jitter in range [-0.5, 0.5]
-                    # WRONG IDEA because the mask_not_already_tested will not updated at the good place
-                    #~ if shift !=0:
-                        #~ jitter = jitter + shift
-                        #~ left_ind = left_ind + shift
-                    
-                    # security to not be outside the fifo
-                    if np.abs(shift) >self.maximum_jitter_shift:
-                        label = LABEL_MAXIMUM_SHIFT
-                    elif (left_ind+shift+self.peak_width)>=self.fifo_residuals.shape[0]:
-                        # normally this should be resolve in the next chunk
-                        label = LABEL_RIGHT_LIMIT
-                    elif (left_ind + shift) < 0:
-                        # TODO assign the previous label ???
-                        label = LABEL_LEFT_LIMIT
-                    else:
-                        label = self.catalogue['cluster_labels'][cluster_idx]
+    def reset_to_not_tested(self, good_spikes):
+        #TODO : more efficient only local !!!!
+        self.local_peaks_mask = self.peakdetector.get_mask_peaks_in_chunk(self.fifo_residuals)
 
-        #security if with jitter the index is out
-        if label>=0:
-            left_ind_check = left_ind - np.round(jitter).astype('int64')
-            if left_ind_check<0:
-                label = LABEL_LEFT_LIMIT
-            elif (left_ind_check+self.peak_width) >=self.fifo_residuals.shape[0]:
-                label = LABEL_RIGHT_LIMIT
-        
-        if label < 0:
-            # set peak tested to not test it again
-            #~ self.mask_not_already_tested[proposed_peak_ind - self.n_span] = False
-            peak_ind = proposed_peak_ind
-
-        #~ self.update_peak_mask(peak_ind, label)
-        #~ t2 = time.perf_counter()
-        #~ print('    update_peak_mask', (t2-t1)*1000)
-        else:
-            # ensure jitter in range [-0.5, 0.5]
-            shift = -int(np.round(jitter))
-            if shift !=0:
-                jitter = jitter + shift
-                left_ind = left_ind + shift
-            
-            peak_ind = left_ind - self.n_left
-        
-        return Spike(peak_ind, label, jitter)
-
+        for spike in good_spikes:
+            peak_ind = spike.index
+            cluster_idx = self.catalogue['label_to_index'][spike.cluster_label]
+            mask = self.sparse_mask[cluster_idx, :]
+            for c in np.nonzero(mask)[0]:
+                self.mask_not_already_tested[peak_ind - self.peak_width - self.n_span:peak_ind + self.peak_width - self.n_span, c] = True
+    
+    def get_no_label_peaks(self):
+        nolabel_indexes, chan_indexes = np.nonzero(~self.mask_not_already_tested)
+        nolabel_indexes += self.n_span
+        nolabel_indexes = nolabel_indexes[nolabel_indexes<(self.chunksize+self.n_span)]
+        bad_spikes = np.zeros(nolabel_indexes.shape[0], dtype=_dtype_spike)
+        bad_spikes['index'] = nolabel_indexes
+        bad_spikes['cluster_label'] = LABEL_UNCLASSIFIED
+        return bad_spikes
 
     def get_best_template(self, left_ind, chan_ind):
         
@@ -367,11 +134,13 @@ class PeelerEngineGeometry(PeelerEngineClassic):
         #~ print('get_best_template')
         #~ print('chan_ind', chan_ind)
         
-        possibles_cluster_idx, = np.nonzero(self.catalogue['sparse_mask'][:, chan_ind])
+        # TODO 
+        
+        possibles_cluster_idx, = np.nonzero(self.sparse_mask[:, chan_ind])
         #~ print(possibles_cluster_idx)
         #~ print('possibles_cluster_idx', possibles_cluster_idx.size)
         
-        #~ print(self.catalogue['sparse_mask'].shape)
+        #~ print(self.sparse_mask.shape)
         
         #~ fig, ax = plt.subplots()
         #~ ax.plot(self.fifo_residuals[:, chan_ind])
@@ -404,11 +173,11 @@ class PeelerEngineGeometry(PeelerEngineClassic):
         
         elif self.argmin_method == 'pythran':
             s = pythran_tools.pythran_loop_sparse_dist(waveform, 
-                                self.catalogue['centers0'],  self.catalogue['sparse_mask'])
+                                self.catalogue['centers0'],  self.sparse_mask)
             cluster_idx = np.argmin(s)
         
         elif self.argmin_method == 'numba':
-            s = numba_loop_sparse_dist_with_geometry(waveform, self.catalogue['centers0'],  self.catalogue['sparse_mask'], possibles_cluster_idx, self.channels_adjacency[chan_ind])
+            s = numba_loop_sparse_dist_with_geometry(waveform, self.catalogue['centers0'],  self.sparse_mask, possibles_cluster_idx, self.channels_adjacency[chan_ind])
             cluster_idx = possibles_cluster_idx[np.argmin(s)]
         
         elif self.argmin_method == 'numpy':
@@ -424,16 +193,23 @@ class PeelerEngineGeometry(PeelerEngineClassic):
         
         
         #~ print('cluster_idx', cluster_idx)
+        #~ print('chan_ind', chan_ind)
         
-        #~ plot_chan = self.channels_adjacency[chan_ind].tolist().index(chan_ind)
+        
         #~ fig, ax = plt.subplots()
-        #~ ax.axvline(plot_chan * self.peak_width - self.n_left)
-        mask = self.catalogue['sparse_mask'][cluster_idx, :]
-        #~ mask = self.channels_adjacency[chan_ind]
-        #~ wf = waveform[:, mask]
-        #~ wf0 = self.catalogue['centers0'][cluster_idx, :, :][:, mask]
+        #~ channels = self.channels_adjacency[chan_ind]
+        #~ wf = waveform[:, channels]
+        #~ wf0 = self.catalogue['centers0'][cluster_idx, :, :][:, channels]
+        #~ wf = waveform
+        #~ wf0 = self.catalogue['centers0'][cluster_idx, :, :]
+        
         #~ ax.plot(wf.T.flatten(), color='k')
         #~ ax.plot(wf0.T.flatten(), color='m')
+        
+        #~ plot_chan = channels.tolist().index(chan_ind)
+        #~ plot_chan = chan_ind
+        #~ ax.axvline(plot_chan * self.peak_width - self.n_left)
+        
         #~ plt.show()
 
         
@@ -449,7 +225,7 @@ class PeelerEngineGeometry(PeelerEngineClassic):
             return False
         
         # criteria multi channel
-        mask = self.catalogue['sparse_mask'][cluster_idx]
+        mask = self.sparse_mask[cluster_idx]
         full_wf0 = self.catalogue['centers0'][cluster_idx,: , :][:, mask]
         full_wf1 = self.catalogue['centers1'][cluster_idx,: , :][:, mask]
         full_wf2 = self.catalogue['centers2'][cluster_idx,: , :][:, mask]
@@ -468,10 +244,16 @@ class PeelerEngineGeometry(PeelerEngineClassic):
         
         # criteria per channel
         crietria_weighted = (wf_nrj>residual_nrj).astype('float') * weight
-        #~ accept_template = np.sum(crietria_weighted) >= 0.9 * np.sum(weight)
-        accept_template = np.sum(crietria_weighted) >= 0.7 * np.sum(weight)
+        accept_template = np.sum(crietria_weighted) >= 0.9 * np.sum(weight)
+        #~ accept_template = np.sum(crietria_weighted) >= 0.7 * np.sum(weight)
 
         #~ if True:
+            
+            #~ max_chan_ind = self.catalogue['clusters'][cluster_idx]['max_on_channel']
+            #~ fig, ax = plt.subplots()
+            #~ ax.plot(self.fifo_residuals[:, max_chan_ind])
+            
+            #~ ax.scatter([left_ind-self.n_left], [self.fifo_residuals[left_ind-self.n_left, max_chan_ind]], color='r')
             
             #~ fig, axs = plt.subplots(nrows=2, sharex=True)
             #~ axs[0].plot(full_wf.T.flatten(), color='b')
@@ -503,6 +285,7 @@ class PeelerEngineGeometry(PeelerEngineClassic):
         
         
         return accept_template
+
